+++ packages/manager/apps/web-domains/src/common/utils/test.provider.tsx
@@ -49,7 +49,7 @@ export const mockShellContext = ({
     },
     getUser: () => ({ ovhSubsidiary: 'FR' }),
     getUserLocale: () => ({}),
-    getRegion: () => ({}),
+    getRegion: () => 'EU',
   },
   shell: {
     navigation: {