+++ packages/manager/apps/web-domains/src/domain/components/DatagridColumns/Domain/DatagridColumnActions.tsx
@@ -1,5 +1,5 @@
 import { NAMESPACES } from '@ovh-ux/manager-common-translations';
-import { ActionMenu, ActionMenuItemProps, BUTTON_COLOR } from '@ovh-ux/muk';
+import { ActionMenu, ActionMenuItemProps } from '@ovh-ux/muk';
 import { useNavigation } from '@ovh-ux/manager-react-shell-client';
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -15,6 +15,7 @@ import { useGenerateUrl } from '@/common/hooks/generateUrl/useGenerateUrl';
 import { useNavigate } from 'react-router-dom';
 import { DomainStateEnum } from '@/domain/enum/domainState.enum';
 import { goToUpdateRenewFrequencyParams } from '@/domain/utils/helpers';
+import { BUTTON_COLOR } from '@ovhcloud/ods-react';
 
 interface DatagridColumnActionsProps {
   readonly serviceName: string;