+++ packages/manager/apps/web-domains/src/domain/hooks/data/query.ts
@@ -38,6 +38,7 @@ import { getOrderCatalog } from '@/domain/data/api/order';
 import {
   getDomainContact,
   getMXPlan,
+  getRedirectionEmail,
   getZimbra,
   updateServiceOption,
 } from '@/common/data/api/common.api';
@@ -71,18 +72,13 @@ export const useGetDomainResource = (serviceName: string) => {
 export const useGetDomainZone = (
   serviceName: string,
   domainResource: TDomainResource,
+  enabled: boolean = false,
 ) => {
-  const isInternalDnsConfiguration =
-    domainResource?.currentState?.dnsConfiguration.configurationType !==
-      DnsConfigurationTypeEnum.EXTERNAL &&
-    domainResource?.currentState?.dnsConfiguration.configurationType !==
-      DnsConfigurationTypeEnum.MIXED;
-
   const { data, isLoading, error } = useQuery<TDomainZone>({
     queryKey: ['domain', 'zone', serviceName],
     queryFn: () => getDomainZone(serviceName),
     retry: false,
-    enabled: isInternalDnsConfiguration,
+    enabled: enabled,
   });
   return {
     domainZone: data,
@@ -266,8 +262,22 @@ export function useEmailService(serviceName: string) {
         // MXplan not found return redirect
       }
 
+      try {
+        const redirection = await getRedirectionEmail(serviceName);
+        if (redirection) {
+          if (redirection.length > 0) {
+            return {
+              serviceDetected: AssociatedEmailsServicesEnum.REDIRECTION,
+              data: redirection[0],
+            };
+          }
+        }
+      } catch (_) {
+        // Redirection not found return nothing
+      }
+
       return {
-        serviceDetected: AssociatedEmailsServicesEnum.REDIRECTION,
+        serviceDetected: AssociatedEmailsServicesEnum.NOTHING,
         data: serviceName,
       };
     },