+++ packages/manager/apps/web-domains/src/domain/pages/domainTabs/dns/dnsModify.tsx
@@ -45,6 +45,7 @@ export default function DnsModifyPage() {
   const { domainZone, isFetchingDomainZone } = useGetDomainZone(
     serviceName,
     domainResource,
+    true,
   );
   const context = useContext(ShellContext);
   const { ovhSubsidiary } = context.environment.getUser();