+++ packages/manager/apps/web-office/src/pages/onboarding/Onboarding.page.tsx
@@ -5,6 +5,7 @@ import { useTranslation } from 'react-i18next';
 import { TEXT_PRESET, Text } from '@ovhcloud/ods-react';
 
 import { NAMESPACES } from '@ovh-ux/manager-common-translations';
+import { useFeatureAvailability } from '@ovh-ux/manager-module-common-api';
 import {
   ButtonType,
   PageLocation,
@@ -23,6 +24,7 @@ export default function Onboarding() {
   const { trackClick } = useOvhTracking();
   const context = useContext(ShellContext);
   const { ovhSubsidiary } = context.environment.getUser();
+  const { data: availability } = useFeatureAvailability(['web-office:order']);
 
   return (
     <OnboardingLayout
@@ -37,21 +39,25 @@ export default function Onboarding() {
         </Text>
       }
       orderButtonLabel={t(`${NAMESPACES.ACTIONS}:order`)}
-      onOrderButtonClick={() => {
-        trackClick({
-          location: PageLocation.page,
-          buttonType: ButtonType.button,
-          actionType: 'action',
-          actions: [ORDER_OFFICE],
-        });
+      onOrderButtonClick={
+        availability?.['web-office:order']
+          ? () => {
+              trackClick({
+                location: PageLocation.page,
+                buttonType: ButtonType.button,
+                actionType: 'action',
+                actions: [ORDER_OFFICE],
+              });
 
-        window.open(
-          GUIDES_LIST.office_cta_order.url[ovhSubsidiary] ||
-            GUIDES_LIST.office_cta_order.url.DEFAULT,
-          '_blank',
-          'noopener',
-        );
-      }}
+              window.open(
+                GUIDES_LIST.office_cta_order.url[ovhSubsidiary] ||
+                  GUIDES_LIST.office_cta_order.url.DEFAULT,
+                '_blank',
+                'noopener',
+              );
+            }
+          : null
+      }
     >
       {[1, 2, 3].map((value: number) => (
         <LinkCard