+++ packages/manager/apps/zimbra/src/pages/dashboard/emailAccounts/EmailAccountsDatagrid.component.tsx
@@ -2,16 +2,21 @@ import { useCallback, useEffect, useMemo, useState } from 'react';
 
 import { useLocation } from 'react-router-dom';
 
+import { RowSelectionState } from '@tanstack/react-table';
 import { useTranslation } from 'react-i18next';
 
-import { ODS_TEXT_PRESET } from '@ovhcloud/ods-components';
-import { OdsText } from '@ovhcloud/ods-components/react';
+import { TEXT_PRESET, Text } from '@ovhcloud/ods-react';
 
 import { NAMESPACES } from '@ovh-ux/manager-common-translations';
-import { Datagrid, DatagridColumn, useBytes } from '@ovh-ux/manager-react-components';
+import { Datagrid, DatagridColumn, useBytes } from '@ovh-ux/muk';
 
 import { AccountStatusBadge, BillingStateBadge, LabelChip } from '@/components';
-import { AccountType, ResourceStatus, getZimbraPlatformListQueryKey } from '@/data/api';
+import {
+  AccountType,
+  ResourceStatus,
+  SlotService,
+  getZimbraPlatformListQueryKey,
+} from '@/data/api';
 import { useAccounts, useSlotServices } from '@/data/hooks';
 import { useDebouncedValue, useOverridePage } from '@/hooks';
 import queryClient from '@/queryClient';
@@ -23,7 +28,7 @@ import { EmailAccountItem } from './EmailAccounts.types';
 
 export const EmailAccountsDatagrid = () => {
   const { t } = useTranslation(['accounts', 'common', NAMESPACES.STATUS]);
-  const [rowSelection, setRowSelection] = useState({});
+  const [rowSelection, setRowSelection] = useState<RowSelectionState>({});
   const [selectedRows, setSelectedRows] = useState([]);
   const [hasADeletingAccount, setHasADeletingAccount] = useState(false);
   const isOverridedPage = useOverridePage();
@@ -102,46 +107,54 @@ export const EmailAccountsDatagrid = () => {
     [],
   );
 
+  useEffect(() => {
+    setSelectedRows(items?.filter((item) => rowSelection[item.id]));
+  }, [items, rowSelection]);
+
   const columns: DatagridColumn<EmailAccountItem>[] = useMemo(
     () => [
       {
         id: 'email_account',
-        cell: (item) => <OdsText preset={ODS_TEXT_PRESET.paragraph}>{item.email}</OdsText>,
+        accessorKey: 'email',
         label: 'common:email_account',
         isSearchable: true,
       },
       {
         id: 'organization',
-        cell: (item) => <LabelChip id={item.organizationId}>{item.organizationLabel}</LabelChip>,
+        accessorKey: 'organizationLabel',
+        cell: ({ getValue }) => <LabelChip id={getValue<string>()}>{getValue<string>()}</LabelChip>,
         label: 'common:organization',
       },
       {
         id: 'offer',
-        cell: (item) => <OdsText preset={ODS_TEXT_PRESET.paragraph}>{item.offer}</OdsText>,
+        accessorKey: 'offer',
         label: 'zimbra_account_datagrid_offer_label',
       },
       {
         id: 'quota',
-        cell: (item) => (
-          <OdsText preset={ODS_TEXT_PRESET.paragraph}>
-            {formatBytes(item.available, 0, 1024)}
-          </OdsText>
+        accessorKey: 'available',
+        cell: ({ getValue }) => (
+          <Text preset={TEXT_PRESET.paragraph}>{formatBytes(getValue<number>(), 0, 1024)}</Text>
         ),
         label: 'zimbra_account_datagrid_quota',
       },
       {
         id: 'status',
-        cell: (item) => <AccountStatusBadge {...item} />,
+        accessorKey: 'email',
+        cell: ({ row }) => <AccountStatusBadge {...row.original} />,
         label: `${NAMESPACES.STATUS}:status`,
       },
       {
         id: 'renewal_type',
-        cell: (item) => <BillingStateBadge service={item.service} />,
+        accessorKey: 'service',
+        cell: ({ getValue }) => <BillingStateBadge service={getValue<SlotService>()} />,
         label: 'zimbra_account_datagrid_renewal_type',
       },
       {
         id: 'tooltip',
-        cell: (item: EmailAccountItem) => <ActionButtonEmailAccount item={item} />,
+        maxSize: 50,
+        accessorKey: '',
+        cell: ({ row }) => <ActionButtonEmailAccount item={row.original} />,
         label: '',
       },
     ],
@@ -160,14 +173,13 @@ export const EmailAccountsDatagrid = () => {
         rowSelection,
         setRowSelection,
         enableRowSelection: ({ original: item }) => isRowSelectable(item),
-        onRowSelectionChange: setSelectedRows,
       }}
       columns={columns.map((column) => ({
         ...column,
-        label: t(column.label),
+        header: t(column.label),
       }))}
-      items={items}
-      totalItems={items.length}
+      data={items}
+      totalCount={items.length}
       hasNextPage={hasNextPage}
       onFetchNextPage={fetchNextPage}
       onFetchAllPages={fetchAllPages}